+++ src/lcm.cpp
@@ -1,5 +1,6 @@
 //@author Erik Edwards
-//@date 2019-2020
+//@date 2018-present
+//@license BSD 3-clause
 
 
 #include <iostream>
@@ -10,7 +11,7 @@
 #include <valarray>
 #include <unordered_map>
 #include <argtable2.h>
-#include "/home/erik/codee/util/cmli.hpp"
+#include "../util/cmli.hpp"
 #include <numeric>
 
 #ifdef I
@@ -28,8 +29,8 @@ int main(int argc, char *argv[])
     const string errstr = ": \033[1;31merror:\033[0m ";
     const string warstr = ": \033[1;35mwarning:\033[0m ";
     const string progstr(__FILE__,string(__FILE__).find_last_of("/")+1,strlen(__FILE__)-string(__FILE__).find_last_of("/")-5);
-    const valarray<uint8_t> oktypes = {8,9,16,17,32,33,64,65};
-    const size_t I = 2, O = 1;
+    const valarray<size_t> oktypes = {8u,9u,16u,17u,32u,33u,64u,65u};
+    const size_t I = 2u, O = 1u;
     ifstream ifs1, ifs2; ofstream ofs1;
     int8_t stdi1, stdi2, stdo1, wo1;
     ioinfo i1, i2, o1;
@@ -97,7 +98,7 @@ int main(int argc, char *argv[])
     if ((i1.T==oktypes).sum()==0 || (i2.T==oktypes).sum()==0)
     {
         cerr << progstr+": " << __LINE__ << errstr << "input data type must be in " << "{";
-        for (auto o : oktypes) { cerr << int(o) << ((o==oktypes[oktypes.size()-1]) ? "}" : ","); }
+        for (auto o : oktypes) { cerr << int(o) << ((o==oktypes[oktypes.size()-1u]) ? "}" : ","); }
         cerr << endl; return 1;
     }
 
@@ -151,7 +152,7 @@ int main(int argc, char *argv[])
     
 
     //Process
-    if (i1.T==8)
+    if (i1.T==8u)
     {
         valarray<int8_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -169,7 +170,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==9)
+    else if (i1.T==9u)
     {
         valarray<uint8_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -187,7 +188,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==16)
+    else if (i1.T==16u)
     {
         valarray<int16_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -205,7 +206,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==17)
+    else if (i1.T==17u)
     {
         valarray<uint16_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -223,7 +224,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==32)
+    else if (i1.T==32u)
     {
         valarray<int32_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -241,7 +242,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==33)
+    else if (i1.T==33u)
     {
         valarray<uint32_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -259,7 +260,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==64)
+    else if (i1.T==64u)
     {
         valarray<int64_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
@@ -277,7 +278,7 @@ int main(int argc, char *argv[])
             catch (...) { cerr << progstr+": " << __LINE__ << errstr << "problem writing output file (Y)" << endl; return 1; }
         }
     }
-    else if (i1.T==65)
+    else if (i1.T==65u)
     {
         valarray<uint64_t> X1(i1.N()), X2(i2.N()), Y(o1.N());
         try { ifs1.read(reinterpret_cast<char*>(&X1[0]),i1.nbytes()); }
