+++ src/size.cpp
@@ -1,5 +1,6 @@
 //@author Erik Edwards
-//@date 2019-2020
+//@date 2018-present
+//@license BSD 3-clause
 
 
 #include <iostream>
@@ -10,7 +11,7 @@
 #include <valarray>
 #include <unordered_map>
 #include <argtable2.h>
-#include "/home/erik/codee/util/cmli.hpp"
+#include "../util/cmli.hpp"
 
 #ifdef I
 #undef I
@@ -27,8 +28,8 @@ int main(int argc, char *argv[])
     const string errstr = ": \033[1;31merror:\033[0m ";
     const string warstr = ": \033[1;35mwarning:\033[0m ";
     const string progstr(__FILE__,string(__FILE__).find_last_of("/")+1,strlen(__FILE__)-string(__FILE__).find_last_of("/")-5);
-    const valarray<uint8_t> oktypes = {1,2,3,8,9,10,16,17,32,33,64,65,101,102,103};
-    const size_t I = 1, O = 1;
+    const valarray<size_t> oktypes = {1u,2u,3u,8u,9u,10u,16u,17u,32u,33u,64u,65u,101u,102u,103u};
+    const size_t I = 1u, O = 1u;
     ifstream ifs1; ofstream ofs1;
     int8_t stdi1, stdo1, wo1;
     ioinfo i1, o1;
@@ -89,7 +90,7 @@ int main(int argc, char *argv[])
     if ((i1.T==oktypes).sum()==0)
     {
         cerr << progstr+": " << __LINE__ << errstr << "input data type must be in " << "{";
-        for (auto o : oktypes) { cerr << int(o) << ((o==oktypes[oktypes.size()-1]) ? "}" : ","); }
+        for (auto o : oktypes) { cerr << int(o) << ((o==oktypes[oktypes.size()-1u]) ? "}" : ","); }
         cerr << endl; return 1;
     }
 
@@ -99,7 +100,7 @@ int main(int argc, char *argv[])
 
     //Set output header info
     o1.F = i1.F;
-    o1.T = (sizeof(size_t)==32) ? 33 : 65;
+    o1.T = (sizeof(size_t)==32u) ? 33u : 65u;
     o1.R = o1.S = o1.H = 1u;
     o1.C = 4u;
 
@@ -121,7 +122,7 @@ int main(int argc, char *argv[])
     
 
     //Process
-    if (i1.T==1)
+    if (i1.T==1u)
     {
     }
     else if (i1.T==2)
@@ -130,40 +131,40 @@ int main(int argc, char *argv[])
     else if (i1.T==3)
     {
     }
-    else if (i1.T==8)
+    else if (i1.T==8u)
     {
     }
-    else if (i1.T==9)
+    else if (i1.T==9u)
     {
     }
-    else if (i1.T==10)
+    else if (i1.T==10u)
     {
     }
-    else if (i1.T==16)
+    else if (i1.T==16u)
     {
     }
-    else if (i1.T==17)
+    else if (i1.T==17u)
     {
     }
-    else if (i1.T==32)
+    else if (i1.T==32u)
     {
     }
-    else if (i1.T==33)
+    else if (i1.T==33u)
     {
     }
-    else if (i1.T==64)
+    else if (i1.T==64u)
     {
     }
-    else if (i1.T==65)
+    else if (i1.T==65u)
     {
     }
-    else if (i1.T==101)
+    else if (i1.T==101u)
     {
     }
-    else if (i1.T==102)
+    else if (i1.T==102u)
     {
     }
-    else if (i1.T==103)
+    else if (i1.T==103u)
     {
     }
     else
