+++ srci/col.cpp
@@ -2,9 +2,9 @@
 #include <complex>
 
 //Declarations
-const valarray<uint8_t> oktypes = {1,2,3,8,9,16,17,32,33,64,65,101,102,103};
-const size_t I = 1, O = 1;
-uint32_t c;
+const valarray<size_t> oktypes = {1u,2u,3u,8u,9u,16u,17u,32u,33u,64u,65u,101u,102u,103u};
+const size_t I = 1u, O = 1u;
+size_t c;
 
 //Description
 string descr;
@@ -27,7 +27,7 @@ struct arg_file  *a_fo = arg_filen("o","ofile","<file>",0,O,"output file (Y)");
 //Get c
 if (a_c->count==0) { c = 0u; }
 else if (a_c->ival[0]<0) { cerr << progstr+": " << __LINE__ << errstr << "c must be nonnegative" << endl; return 1; }
-else { c = uint32_t(a_c->ival[0]); }
+else { c = size_t(a_c->ival[0]); }
 if (c>=i1.C) { cerr << progstr+": " << __LINE__ << errstr << "c must be int in [0 C-1]" << endl; return 1; }
 
 //Checks
@@ -40,7 +40,7 @@ o1.S = i1.S; o1.H = i1.H;
 //Other prep
 
 //Process
-if (i1.T==1)
+if (i1.T==1u)
 {
     valarray<float> X(i1.N()), Y(o1.N());
     try { ifs1.read(reinterpret_cast<char*>(&X[0]),i1.nbytes()); }
