+++ srci/row.cpp
@@ -2,9 +2,9 @@
 #include <complex>
 
 //Declarations
-const valarray<uint8_t> oktypes = {1,2,3,8,9,16,17,32,33,64,65,101,102,103};
-const size_t I = 1, O = 1;
-uint32_t r;
+const valarray<size_t> oktypes = {1u,2u,3u,8u,9u,16u,17u,32u,33u,64u,65u,101u,102u,103u};
+const size_t I = 1u, O = 1u;
+size_t r;
 
 //Description
 string descr;
@@ -25,19 +25,20 @@ struct arg_file  *a_fo = arg_filen("o","ofile","<file>",0,O,"output file (Y)");
 //Get r
 if (a_r->count==0) { r = 0u; }
 else if (a_r->ival[0]<0) { cerr << progstr+": " << __LINE__ << errstr << "r must be nonnegative" << endl; return 1; }
-else { r = uint32_t(a_r->ival[0]); }
+else { r = size_t(a_r->ival[0]); }
 if (r>=i1.R) { cerr << progstr+": " << __LINE__ << errstr << "r must be int in [0 R-1]" << endl; return 1; }
 
 //Checks
 
 //Set output header info
 o1.F = i1.F; o1.T = i1.T;
-o1.R = 1u; o1.C = i1.C; o1.S = i1.S; o1.H = i1.H;
+o1.R = 1u; o1.C = i1.C;
+o1.S = i1.S; o1.H = i1.H;
 
 //Other prep
 
 //Process
-if (i1.T==1)
+if (i1.T==1u)
 {
     valarray<float> X(i1.N()), Y(o1.N());
     try { ifs1.read(reinterpret_cast<char*>(&X[0]),i1.nbytes()); }
